rails 4 doesn't generate public/index.html

+++ test_support/bin/test.sh
@@ -3,7 +3,7 @@
 # Create a default rails appliaction, install blacklight, and run all the tests.
 
 #If null or empty, use default value
-RAILS_VERSION=${RAILS_VERSION:-"~> 3.2"}
+RAILS_VERSION=${RAILS_VERSION:-"4.0.0.beta1"}
 JETTY_URL=${JETTY_URL:-"https://github.com/projectblacklight/blacklight-jetty/zipball/v4.0.0"}
 
 before="$(date +%s)"
@@ -69,7 +69,10 @@ fi
 rails new test_app
 
 cd test_app
-rm public/index.html
+
+if [ -e public/index.html ]; then
+  rm public/index.html
+fi
 
 echo "
 platforms :jruby do
